pyproject.toml
[project]
name = "clinvar_gk_pilot"
authors = [{ name = "Kyle Ferriter", email = "kferrite@broadinstitute.org" }]
description = "Conversion of ClinVar Variation expressions to GA4GH GKS forms"
readme = "README.md"
requires-python = ">=3.11"
keywords = ["bioinformatics", "clinvar"]
license = { text = "MIT" }
classifiers = ["Programming Language :: Python :: 3"]
dependencies = [
    "google-cloud-storage~=2.13.0",
    #"ga4gh.vrs[extras] @ git+https://github.com/ga4gh/vrs-python@2.0.0-a6",
    "ga4gh.vrs[extras] @ git+https://github.com/ga4gh/vrs-python@main-draft"
]
dynamic = ["version"]

[project.optional-dependencies]
dev = [
    "ipykernel",
    "black~=23.9.1",
    "isort~=5.12.0",
    "ruff~=0.1.5",
    "pytest~=7.4.3",
    "pylint~=3.0.2",
]

[project.scripts]
clinvar-gk-pilot = "clinvar_gk_pilot.main:main"

[tool.setuptools.packages.find]
include = ["clinvar_gk_pilot*"]

# [tool.setuptools.package-data]
# "clinvar_ingest" = ["*.json", ".*.env"]
# "clinvar_ingest.cloud.bigquery.bq_json_schemas" = ["*.json"]

[tool.isort]
profile = "black"
